% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmleFH.R
\name{mmleFH}
\alias{mmleFH}
\title{Marginal MLEs for the Fay-Herriot model}
\usage{
mmleFH(y, X, V, pSS = 0, pdf = 0)
}
\arguments{
\item{y}{direct data follwing normal model \eqn{y\sim N(\theta,V\sigma^2)}}

\item{X}{linking model predictors \eqn{ \theta\sim N(X\beta,\tau^2 I)}}

\item{V}{covariance matrix to scale}

\item{pSS}{prior sum of squares for estimate of \eqn{\sigma^2}}

\item{pdf}{prior degrees of freedom for estimate of \eqn{\sigma^2}}
}
\description{
Marginal MLEs for the Fay-Herriot random effects model where 
the covariance matrix for the sampling model is known to scale.
}
