% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmleFHP.R
\name{mmleFHP}
\alias{mmleFHP}
\title{Marginal MLEs for the Fay-Herriot model with known covariance}
\usage{
mmleFHP(y, X, Sigma)
}
\arguments{
\item{y}{direct data follwing normal model \eqn{y\sim N(\theta,\Sigma)}}

\item{X}{linking model predictors \eqn{ \theta\sim N(X\beta,\tau^2 I)}}

\item{Sigma}{covariance matrix in sampling model}
}
\description{
Marginal MLEs for the Fay-Herriot random effects model where 
the covariance matrix for the sampling model is known
}
