% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{qr.lmFAB}
\alias{qr.lmFAB}
\title{QR decomposition}
\usage{
\method{qr}{lmFAB}(x, ...)
}
\arguments{
\item{x}{\code{lmFAB} object}

\item{...}{see \code{qr.lm}, if you can find it}
}
\value{
qr decomposition for a design matrix
}
\description{
QR decomposition for lmFAB objects
}
