% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{summary.glmFAB}
\alias{summary.glmFAB}
\title{Summarizing Generalized Linear Model Fits with FAB Inference}
\usage{
\method{summary}{glmFAB}(object, dispersion = NULL,
  correlation = FALSE, symbolic.cor = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{glmFAB}}

\item{dispersion}{see \code{summary.glm}}

\item{correlation}{see \code{summary.glm}}

\item{symbolic.cor}{see \code{summary.glm}}

\item{...}{see \code{summary.glm}}
}
\description{
\code{summary} method for class \code{glmFAB}
}
\details{
A mod of \code{summary.glm} that shows FAB p-values in table
}
