% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{summary.lmFAB}
\alias{summary.lmFAB}
\title{Summarizing Linear Model Fits with FAB Inference}
\usage{
\method{summary}{lmFAB}(object, correlation = FALSE,
  symbolic.cor = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{lmFAB}}

\item{correlation}{see \code{summary.lm}}

\item{symbolic.cor}{see \code{summary.lm}}

\item{...}{see \code{summary.lm}}
}
\description{
\code{summary} method for class \code{lmFAB}
}
\details{
A mod of \code{summary.lm} that shows FAB p-values in table
}
